# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Motus API",
        "description": "API REST pour le jeu de devinettes de mots Motus. Authentification par session cookie (Sanctum SPA).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Serveur local de développement"
        }
    ],
    "paths": {
        "/api/games/{game}/attempts": {
            "post": {
                "tags": [
                    "Attempts"
                ],
                "summary": "Soumettre un essai sur une partie",
                "description": "Valide le mot (longueur exacte, lettres autorisées), calcule le scoring par lettre, et met à jour la partie. Renvoie l'essai créé et la partie rafraîchie (le mot cible n'est révélé qu'une fois la partie terminée).",
                "operationId": "fbc80c44f2cbc8ee74591118abc8a60e",
                "parameters": [
                    {
                        "name": "game",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "word"
                                ],
                                "properties": {
                                    "word": {
                                        "description": "Mot proposé. Doit avoir la même longueur que le mot cible et ne contenir que des lettres, traits d'union et apostrophes.",
                                        "type": "string",
                                        "example": "MAISON"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Essai enregistré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AttemptSubmission"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "La partie appartient à un autre utilisateur ou n'est plus en cours"
                    },
                    "422": {
                        "description": "Mot invalide (mauvaise longueur ou caractères non autorisés)"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Créer un compte et démarrer une session",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Alice"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "alice@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123",
                                        "minLength": 8
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Compte créé, session ouverte",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Se connecter et démarrer une session",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "alice@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Session ouverte",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Identifiants incorrects"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Invalider la session courante",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "204": {
                        "description": "Session invalidée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Récupérer l'utilisateur courant",
                "operationId": "e5478e84ace512f8d9653dd7e920ea2a",
                "responses": {
                    "200": {
                        "description": "Utilisateur courant",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Alice"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "alice@example.com"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Attempt": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 42
                    },
                    "attempt_number": {
                        "type": "integer",
                        "example": 1
                    },
                    "attempted_word": {
                        "type": "string",
                        "example": "MAISON"
                    },
                    "result": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "enum": [
                                "correct",
                                "present",
                                "absent"
                            ]
                        }
                    },
                    "is_correct": {
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Game": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 7
                    },
                    "difficulty": {
                        "type": "string",
                        "example": "easy",
                        "nullable": true,
                        "enum": [
                            "easy",
                            "medium",
                            "hard"
                        ]
                    },
                    "length": {
                        "type": "integer",
                        "example": 5,
                        "nullable": true
                    },
                    "first_letter": {
                        "type": "string",
                        "example": "M",
                        "nullable": true
                    },
                    "attempts_count": {
                        "type": "integer",
                        "example": 2
                    },
                    "max_attempts": {
                        "type": "integer",
                        "example": 6
                    },
                    "status": {
                        "type": "string",
                        "example": "in_progress",
                        "enum": [
                            "in_progress",
                            "won",
                            "lost"
                        ]
                    },
                    "score": {
                        "type": "integer",
                        "example": 0
                    },
                    "completed_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "attempts": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Attempt"
                        }
                    },
                    "word": {
                        "description": "Le mot cible. Null tant que la partie n'est pas terminée, révélé quand la partie est gagnée ou perdue.",
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "AttemptSubmission": {
                "properties": {
                    "attempt": {
                        "$ref": "#/components/schemas/Attempt"
                    },
                    "game": {
                        "$ref": "#/components/schemas/Game"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "name": "X-XSRF-TOKEN",
                "in": "header"
            }
        }
    },
...
}